refactor invite params to custom hook
