+++ apps/intake/src/pages/ReviewPaperwork.tsx
@@ -16,13 +16,7 @@ import {
   usePaperworkContext,
 } from 'ui-components';
 import { ZambdaClient, useUCZambdaClient } from 'ui-components/lib/hooks/useUCZambdaClient';
-import {
-  InviteParticipantRequestParameters,
-  VisitType,
-  makeValidationSchema,
-  pickFirstValueFromAnswerItem,
-  uuidRegex,
-} from 'utils';
+import { VisitType, makeValidationSchema, pickFirstValueFromAnswerItem, uuidRegex } from 'utils';
 import { ValidationError } from 'yup';
 import { otherColors } from '../IntakeThemeProvider';
 import api from '../api/zapehrApi';
@@ -32,6 +26,7 @@ import { UNEXPECTED_ERROR_CONFIG } from '../helpers';
 import { getLocaleDateTimeString } from '../helpers/dateUtils';
 import useAppointmentNotFoundInformation from '../helpers/information';
 import { useGetFullName } from '../hooks/useGetFullName';
+import { usePaperworkInviteParams } from '../hooks/usePaperworkInviteParams';
 import { useSetLastActiveTime } from '../hooks/useSetLastActiveTime';
 import i18n from '../lib/i18n';
 import { useCreateInviteMutation } from '../telemed/features/waiting-room';
@@ -106,40 +101,7 @@ const ReviewPaperwork = (): JSX.Element => {
 
   const createInviteMutation = useCreateInviteMutation();
 
-  const inviteParams = useMemo((): null | Omit<InviteParticipantRequestParameters, 'appointmentId'> => {
-    const page = completedPaperwork.find((page) => page.linkId === 'invite-participant-page');
-
-    if (!page) {
-      // in-person doesn't need invite participant
-      return null;
-    }
-
-    const answers: { [key: string]: string | undefined } = {};
-
-    page.item?.forEach((item) => {
-      const linkId = item.linkId;
-      let value: string | undefined;
-      if (item?.answer?.[0]?.valueString) {
-        value = item.answer[0].valueString;
-      }
-      answers[linkId] = value;
-    });
-
-    if (answers['invite-from-another-device'] !== 'Yes, I will add invite details below') {
-      return null;
-    }
-
-    if (answers['invite-phone']) {
-      answers['invite-phone'] = answers['invite-phone'].replace(/\D/g, '');
-    }
-
-    return {
-      firstName: answers['invite-first']!,
-      lastName: answers['invite-last']!,
-      emailAddress: answers['invite-email']!,
-      phoneNumber: answers['invite-phone']!,
-    };
-  }, [completedPaperwork]);
+  const inviteParams = usePaperworkInviteParams(completedPaperwork);
 
   const navigateToWaitingRoom = useCallback(
     (error: boolean): void => {
